riscv: cheri: Make sbi_ecall() Zcheri compatible
Improve code portibility to support Zcheri extension by using
assembly macros and change the data type presenting general
purpose register as uintptr_t.

+++ arch/riscv/include/asm/sbi.h
@@ -143,10 +143,10 @@ struct sbiret {
 	long value;
 };
 
-struct sbiret sbi_ecall(int ext, int fid, unsigned long arg0,
-			unsigned long arg1, unsigned long arg2,
-			unsigned long arg3, unsigned long arg4,
-			unsigned long arg5);
+struct sbiret sbi_ecall(int ext, int fid, uintptr_t arg0,
+			uintptr_t arg1, uintptr_t arg2,
+			uintptr_t arg3, uintptr_t arg4,
+			uintptr_t arg5);
 
 #ifdef CONFIG_SBI_V01
 void sbi_console_putchar(int ch);
